Big update that adds menubar actions
- New menu bar

Because of the addition of a menubar, the rendering had to be updated
to run within a thread else it would block the UI thread.

Current Menu Actions:
 - Seed -> Open (open seed file - not implemented)
 - Seed -> Generate (New random seed - implemented)
 - Action -> Pause (Paused rendering - implemented)
 - Action -> Play (Unpause rendering - implemented)
 - Action -> Stop (End simulation - implemented)
 - Action -> Start (Start simulation - implemented)

Atm, Action -> Start is not useful since generating a new seed will
automatically start.

To support threading, the actual cycling of life updates is handled by
an implemention of LifeRunner. Currently using ThreadedLifeRunner.

Also removed dependency on java-color-grid and implemented similar code
right in the project.